address review: drop the adapter and the checksum indirection
PackageVerifier only existed to hold a results map nothing ever read,
and a mutex guarding it. Verifier now implements pkg.Verifier directly.

The ParseChecksums re-export is gone: it dated from when the checksum
parser was meant to be swappable, which it no longer is. The library
already reports which files a signature covers, so ErrNameMismatch wraps
that rather than rebuilding the list.

Drops the unused TrustDir and EnsureTrustDir, left over from the
deferred pkg trust command, and two tests that exercised go-signify
through plakar rather than the behaviour here.
