+++ appcontext/appcontext.go
@@ -15,10 +15,10 @@ import (
 type AppContext struct {
 	*kcontext.KContext
 
-	cookies     *cookies.Manager         `msgpack:"-"`
-	pkgmgr      *pkg.Manager             `msgpack:"-"`
-	pkgverifier *signify.PackageVerifier `msgpack:"-"`
-	Config      *config.Config           `msgpack:"-"`
+	cookies     *cookies.Manager  `msgpack:"-"`
+	pkgmgr      *pkg.Manager      `msgpack:"-"`
+	pkgverifier *signify.Verifier `msgpack:"-"`
+	Config      *config.Config    `msgpack:"-"`
 
 	ConfigDir string
 	secret    []byte
@@ -110,11 +110,11 @@ func (c *AppContext) GetPkgManager() *pkg.Manager {
 	return c.pkgmgr
 }
 
-func (c *AppContext) SetPkgVerifier(verifier *signify.PackageVerifier) {
+func (c *AppContext) SetPkgVerifier(verifier *signify.Verifier) {
 	c.pkgverifier = verifier
 }
 
-func (c *AppContext) GetPkgVerifier() *signify.PackageVerifier {
+func (c *AppContext) GetPkgVerifier() *signify.Verifier {
 	return c.pkgverifier
 }
