+++ go.mod
@@ -12,7 +12,7 @@ require (
 	github.com/PlakarKorp/integrations/stdio v1.1.0
 	github.com/PlakarKorp/integrations/tar v1.1.0
 	github.com/PlakarKorp/kloset v1.2.0-alpha.2.0.20260728134607-9310b71c3adc
-	github.com/PlakarKorp/pkg v1.1.3-0.20260815233227-6ac26c3d50ec
+	github.com/PlakarKorp/pkg v1.1.3-0.20260817121456-e41c09352892
 	github.com/alecthomas/chroma v0.10.0
 	github.com/anacrolix/fuse v0.3.3-0.20260723023734-9e1272bc0085
 	github.com/charmbracelet/bubbles v1.0.0
